pyproject.toml
[project]
name = "app"
version = "0.1.0"
dependencies = [ "flask==3.0.3", "python-dotenv==1.0.1", "requests==2.32.3", "google-generativeai==0.8.3", "pymongo==4.11.1" ]
requires-python = "~=3.12.0"
